Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -4,9 +4,9 @@
 
 int main() {
     float r;
-    scanf("%.2f",r);
+    scanf("%f",r);
     float pie = 3.14;
-    float area = pie*r*r;
-    printf("Area: %.2f",area);
+    double area = pie*r*r;
+    printf("Area: %.2lf",area);
     return 0;
 }
